Solved 2020 day 18 part 1

+++ src/main/java/nl/mvdr/adventofcode/adventofcode2020/day18/Expression.java
@@ -77,5 +77,5 @@ interface Expression {
     }
     
     /** @return the value of this expression */
-    int evaluate();
+    long evaluate();
 }
