+++ src/main/java/nl/mvdr/adventofcode/adventofcode2020/day18/Numeral.java
@@ -18,7 +18,7 @@ record Numeral(int value) implements Expression {
     }
     
     @Override
-    public int evaluate() {
+    public long evaluate() {
         return value;
     }
 }
